Provide default implementation as part of XMLProvider
